#1697 - Part 3 - Extracting Simulation Metrics to it's own panel.
